refactor(supabase): move auth test files to __tests__/ directory
